// src/app/models/pedido.ts
export interface Pedido {
    id?:                  any;
    dataPedido?:           any;
    nomeCliente:       string;     
    status:            string;
    descricao:         string;
    endereco:          string;
    cep:               string;
    cidade:            string;
    rg:                string;
    celular:           string;
    valor:             string;   
    
  } 